Refactor impl into emulator.ts.
