+++ packages-exp/auth-exp/src/core/auth/auth_impl.ts
@@ -272,20 +272,6 @@ export class AuthImpl implements Auth, _FirebaseService {
     this.languageCode = _getUserLanguage();
   }
 
-  _useEmulator(url: string, options?: { disableWarnings: boolean }): void {
-    _assert(this._canInitEmulator, this, AuthErrorCode.EMULATOR_CONFIG_FAILED);
-
-    _assert(
-      /^https?:\/\//.test(url),
-      this,
-      AuthErrorCode.INVALID_EMULATOR_SCHEME
-    );
-
-    this.config.emulator = { url };
-    this.settings.appVerificationDisabledForTesting = true;
-    emitEmulatorWarning(!!options?.disableWarnings);
-  }
-
   async _delete(): Promise<void> {
     this._deleted = true;
   }
@@ -568,45 +554,5 @@ class Subscription<T> {
   get next(): NextFn<T | null> {
     _assert(this.observer, this.auth, AuthErrorCode.INTERNAL_ERROR);
     return this.observer.next.bind(this.observer);
-  }
-}
-
-function emitEmulatorWarning(disableBanner: boolean): void {
-  function attachBanner(): void {
-    const el = document.createElement('p');
-    const sty = el.style;
-    el.innerText =
-      'Running in emulator mode. Do not use with production credentials.';
-    sty.position = 'fixed';
-    sty.width = '100%';
-    sty.backgroundColor = '#ffffff';
-    sty.border = '.1em solid #000000';
-    sty.color = '#ff0000';
-    sty.bottom = '0px';
-    sty.left = '0px';
-    sty.margin = '0px';
-    sty.zIndex = '10000';
-    sty.textAlign = 'center';
-    el.classList.add('firebase-emulator-warning');
-    document.body.appendChild(el);
-  }
-
-  if (typeof console !== 'undefined' && typeof console.info === 'function') {
-    console.info(
-      'WARNING: You are using the Auth Emulator,' +
-        ' which is intended for local testing only.  Do not use with' +
-        ' production credentials.'
-    );
-  }
-  if (
-    typeof window !== 'undefined' &&
-    typeof document !== 'undefined' &&
-    !disableBanner
-  ) {
-    if (document.readyState === 'loading') {
-      window.addEventListener('DOMContentLoaded', attachBanner);
-    } else {
-      attachBanner();
-    }
   }
 }
